Added function for setting up game state
