+++ TictactoeVer2.csproj
@@ -49,7 +49,7 @@
     <Compile Include="PlayerModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tictactoe.cs" />
-    <Compile Include="UserInputParser.cs" />
+    <Compile Include="UserInputHandler.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
